+++ prisonapp/src/main/java/com/prisonapp/business/controller/dw_test/TestController.java
@@ -1,0 +1,18 @@
+package com.prisonapp.business.controller.dw_test;
+
+import io.swagger.annotations.Api;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@Api("测试接口")
+@RestController
+@RequestMapping("/Test")
+public class TestController {
+// 我现在用hashmap实现的 它那个有九种情况 你调用成功过了码还没有
+    public void Test(){
+
+    }
+
+
+
+}
