--- prisonapp/src/main/java/com/prisonapp/business/controller/test/TestController.java
@@ -1,24 +1,0 @@
-package com.prisonapp.business.controller.test;
-
-import com.prisonapp.model.LoadModel;
-import com.prisonapp.model.MessagePush;
-import com.prisonapp.tool.HttpURLConnectionDemo;
-import com.prisonapp.tool.JsonUtils;
-import io.swagger.annotations.Api;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.util.HashMap;
-
-@Api("测试接口")
-@RestController
-@RequestMapping("/Test")
-public class TestController {
-// 我现在用hashmap实现的 它那个有九种情况 你调用成功过了码还没有
-    public void Test(){
-
-    }
-
-
-
-}
