+++ prisonapp/src/main/java/com/prisonapp/business/dao/dw_user/UserDao.java
@@ -1,0 +1,11 @@
+package com.prisonapp.business.dao.dw_user;
+
+import com.prisonapp.business.entity.dw_user.UserModel;
+import org.apache.ibatis.annotations.Mapper;
+import org.apache.ibatis.annotations.Param;
+
+@Mapper
+public interface UserDao {
+
+    UserModel login(@Param("userName") String userName);
+}
