+++ prisonapp/src/main/java/com/prisonapp/business/entity/dw_message/NotificationMessageModel.java
@@ -1,0 +1,69 @@
+package com.prisonapp.business.entity.dw_message;
+
+import java.sql.Timestamp;
+//保外人员的通知列表
+public class NotificationMessageModel {
+    public String type;
+    public String typeName;
+    public String content;
+    public java.sql.Timestamp timestamp;
+    public long detailType;
+    public String detailTypeName;
+    public int unreadCount;
+
+    public String getType() {
+        return type;
+    }
+
+    public void setType(String type) {
+        this.type = type;
+    }
+
+    public String getTypeName() {
+        return typeName;
+    }
+
+    public void setTypeName(String typeName) {
+        this.typeName = typeName;
+    }
+
+    public String getContent() {
+        return content;
+    }
+
+    public void setContent(String content) {
+        this.content = content;
+    }
+
+    public Timestamp getTimestamp() {
+        return timestamp;
+    }
+
+    public void setTimestamp(Timestamp timestamp) {
+        this.timestamp = timestamp;
+    }
+
+    public long getDetailType() {
+        return detailType;
+    }
+
+    public void setDetailType(long detailType) {
+        this.detailType = detailType;
+    }
+
+    public String getDetailTypeName() {
+        return detailTypeName;
+    }
+
+    public void setDetailTypeName(String detailTypeName) {
+        this.detailTypeName = detailTypeName;
+    }
+
+    public int getUnreadCount() {
+        return unreadCount;
+    }
+
+    public void setUnreadCount(int unreadCount) {
+        this.unreadCount = unreadCount;
+    }
+}
