+++ prisonapp/src/main/java/com/prisonapp/business/entity/dw_message/SearchNotificationModel.java
@@ -1,0 +1,67 @@
+package com.prisonapp.business.entity.dw_message;
+
+public class SearchNotificationModel {
+    private String type;
+    private String typeName;
+    private String detailType;
+    private String detailTypeName;
+    private int messageCount;
+    private String content;
+    private String timestamp;
+
+    public String getType() {
+        return type;
+    }
+
+    public void setType(String type) {
+        this.type = type;
+    }
+
+    public String getTypeName() {
+        return typeName;
+    }
+
+    public void setTypeName(String typeName) {
+        this.typeName = typeName;
+    }
+
+    public String getDetailType() {
+        return detailType;
+    }
+
+    public void setDetailType(String detailType) {
+        this.detailType = detailType;
+    }
+
+    public String getDetailTypeName() {
+        return detailTypeName;
+    }
+
+    public void setDetailTypeName(String detailTypeName) {
+        this.detailTypeName = detailTypeName;
+    }
+
+    public int getMessageCount() {
+        return messageCount;
+    }
+
+    public void setMessageCount(int messageCount) {
+        this.messageCount = messageCount;
+    }
+
+    public String getContent() {
+        return content;
+    }
+
+    public void setContent(String content) {
+        this.content = content;
+    }
+
+    public String getTimestamp() {
+        return timestamp;
+    }
+
+    public void setTimestamp(String timestamp) {
+        this.timestamp = timestamp;
+    }
+}
