+++ prisonapp/src/main/java/com/prisonapp/business/entity/dw_supervise/ApplyRecord.java
@@ -1,0 +1,42 @@
+package com.prisonapp.business.entity.dw_supervise;
+
+import java.sql.Timestamp;
+
+public class ApplyRecord {
+    public int statusCode;
+    public String  status;
+    public java.sql.Timestamp timestamp;
+    public String person;
+
+    public int getStatusCode() {
+        return statusCode;
+    }
+
+    public void setStatusCode(int statusCode) {
+        this.statusCode = statusCode;
+    }
+
+    public String getStatus() {
+        return status;
+    }
+
+    public void setStatus(String status) {
+        this.status = status;
+    }
+
+    public Timestamp getTimestamp() {
+        return timestamp;
+    }
+
+    public void setTimestamp(Timestamp timestamp) {
+        this.timestamp = timestamp;
+    }
+
+    public String getPerson() {
+        return person;
+    }
+
+    public void setPerson(String person) {
+        this.person = person;
+    }
+}
