+++ prisonapp/src/main/java/com/prisonapp/token/TokenService.java
@@ -2,7 +2,7 @@ package com.prisonapp.token;
 
 import com.auth0.jwt.JWT;
 import com.auth0.jwt.algorithms.Algorithm;
-import com.prisonapp.business.entity.user.UserModel;
+import com.prisonapp.business.entity.dw_user.UserModel;
 import org.springframework.stereotype.Service;
 
 import java.util.Date;
@@ -19,7 +19,7 @@ public  class TokenService {
      */
     public String getToken(UserModel user) {
         Date start = new Date();
-        long currentTime = System.currentTimeMillis() + 60* 60 * 1000;//一小时有效时间
+        long currentTime = System.currentTimeMillis() +7* 24*60* 60 * 1000;//7天有效时间
         Date end = new Date(currentTime);
         String token = "";
