+++ prisonapp/src/main/resources/mapper/dw_message/MessageMapper.xml
@@ -1,0 +1,39 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+        "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper namespace="com.prisonapp.business.dao.dw_message.MessageDao">
+    <select id="getNotificationList" parameterType="String" resultType="com.prisonapp.business.entity.dw_message.NotificationMessageModel">
+
+      select  modular as type ,modularname as typeName,detailtype as detailType,detailtypename as detailTypeName,content,messagetime as timestamp from  (SELECT * FROM "t_protectionmessage"  where modular='1'  and personid=#{userId}  ORDER BY  messagetime desc  limit  1) as a
+UNION ALL
+select  modular as type ,modularname as typeName,detailtype as detailType,detailtypename as detailTypeName,content,messagetime as timestamp  from  (SELECT * FROM "t_protectionmessage"  where modular='2'  and personid=#{userId}  ORDER BY  messagetime desc limit  1) as a
+UNION ALL
+select  modular as type ,modularname as typeName,detailtype as detailType,detailtypename as detailTypeName,content,messagetime as timestamp  from  (SELECT * FROM "t_protectionmessage"  where modular='3'  and personid=#{userId}  ORDER BY  messagetime desc limit  1) as a
+UNION ALL
+select  modular as type ,modularname as typeName,detailtype as detailType,detailtypename as detailTypeName,content,messagetime as timestamp  from  (SELECT * FROM "t_protectionmessage"  where modular='4'  and personid=#{userId}  ORDER BY  messagetime desc limit  1) as a
+UNION ALL
+select  modular as type ,modularname as typeName,detailtype as detailType,detailtypename as detailTypeName,content,messagetime as timestamp  from  (SELECT * FROM "t_protectionmessage"  where modular='5'  and personid=#{userId}  ORDER BY  messagetime desc limit  1) as a
+    </select>
+    <select id="getMessageList" resultType="com.prisonapp.business.entity.dw_message.MessageListModel">
+        select  id as code,modular as type ,modularname as typeName,content,messagetime as timesta,readmessage as isRead from t_protectionmessage where personid=#{userId} and modular=#{type}  ORDER BY timesta  desc limit #{requestCount}
+    </select>
+    <update id="readMessage" parameterType="String">
+        update t_protectionmessage set readmessage=true where modular=#{type} and personid=#{userId}
+    </update>
+    <select id="searchNotification" resultType="com.prisonapp.business.entity.dw_message.SearchNotificationModel">
+                select * from (SELECT content, modular as type,modularname as typeName,detailtype as detailType,detailtypename as detailTypeName,messagetime as timestamp FROM "t_protectionmessage"  where modular='1'  and  personid=#{userId}  ORDER BY  messagetime desc  limit  1 ) as a where  content like concat('%',#{key},'%')
+	 UNION ALL  select * from (SELECT content, modular as type,modularname as typeName,detailtype as detailType,detailtypename as detailTypeName,messagetime as timestamp FROM "t_protectionmessage"  where modular='2'  and  personid=#{userId}  ORDER BY  messagetime desc  limit  1 ) as a where  content like concat('%',#{key},'%')
+	 UNION ALL	select * from (SELECT content, modular as type,modularname as typeName,detailtype as detailType,detailtypename as detailTypeName,messagetime as timestamp FROM "t_protectionmessage"  where modular='3'  and  personid=#{userId}  ORDER BY  messagetime desc  limit  1 ) as a where  content like concat('%',#{key},'%')
+	 UNION ALL	select * from (SELECT content, modular as type,modularname as typeName,detailtype as detailType,detailtypename as detailTypeName,messagetime as timestamp FROM "t_protectionmessage"  where modular='4'  and  personid=#{userId}  ORDER BY  messagetime desc  limit  1 ) as a where  content like concat('%',#{key},'%')
+	 UNION ALL	select * from (SELECT content, modular as type,modularname as typeName,detailtype as detailType,detailtypename as detailTypeName,messagetime as timestamp FROM "t_protectionmessage"  where modular='5'  and  personid=#{userId}  ORDER BY  messagetime desc  limit  1 ) as a where  content like concat('%',#{key},'%')
+    </select>
+    <select id="unreadCount" resultType="com.prisonapp.business.entity.dw_message.MessageModel">
+        select * from t_protectionmessage where personid=#{userId} and  modular=#{type} and readmessage='f'
+    </select>
+
+    <select id="messageTotalCount" resultType="com.prisonapp.business.entity.dw_message.MessageModel">
+         select * from t_protectionmessage where personid=#{userId} and  modular=#{type}
+    </select>
+
+
+</mapper>
