+++ prisonapp/src/main/resources/mapper/dw_user/UserMapper.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+        "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper namespace="com.prisonapp.business.dao.dw_user.UserDao">
+    <select id="login" resultType="com.prisonapp.business.entity.dw_user.UserModel" parameterType="String">
+        select * from t_user where accountname=#{userName}
+    </select>
+</mapper>
