--- prisonapp/src/main/resources/mapper/user/UserMapper.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
-        "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.prisonapp.business.dao.user.UserDao">
-    <select id="login" resultType="com.prisonapp.business.entity.user.UserModel" parameterType="String">
-        select * from t_user where accountname=#{userName}
-    </select>
-</mapper>
